CASSANDRA-20884 - Use shared JMX classes which were moved to the in-jvm-dtest API
