+++ .build/parent-maven-pom.xml
@@ -524,7 +524,7 @@
       <dependency>
         <groupId>org.apache.cassandra</groupId>
         <artifactId>dtest-api</artifactId>
-        <version>0.0.17</version>
+        <version>0.0.18</version>
         <scope>test</scope>
       </dependency>
       <dependency>
