+++ test/distributed/org/apache/cassandra/distributed/impl/AbstractCluster.java
@@ -228,12 +228,6 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster<I
             withSharedClasses(SHARED_PREDICATE);
         }
 
-        @SuppressWarnings("unchecked")
-        private B self()
-        {
-            return (B) this;
-        }
-
         public B withNodeProvisionStrategy(INodeProvisionStrategy.Factory nodeProvisionStrategy)
         {
             this.nodeProvisionStrategy = nodeProvisionStrategy;
@@ -618,7 +612,7 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster<I
     }
 
     @VisibleForTesting
-    InstanceConfig createInstanceConfig(int nodeNum)
+    public InstanceConfig createInstanceConfig(int nodeNum)
     {
         INodeProvisionStrategy provisionStrategy = nodeProvisionStrategy.create(subnet, portMap);
         Collection<String> tokens = tokenSupplier.tokens(nodeNum);
