+++ test/distributed/org/apache/cassandra/distributed/impl/IsolatedJmx.java
@@ -33,17 +33,18 @@ import javax.management.remote.rmi.RMIConnectorServer;
 import javax.management.remote.rmi.RMIJRMPServerImpl;
 
 import com.google.common.util.concurrent.Uninterruptibles;
+
 import org.slf4j.Logger;
 
 import org.apache.cassandra.config.EncryptionOptions;
 import org.apache.cassandra.config.JMXServerOptions;
 import org.apache.cassandra.distributed.api.IInstance;
 import org.apache.cassandra.distributed.api.IInstanceConfig;
+import org.apache.cassandra.distributed.shared.jmx.RMICloseableClientSocketFactory;
+import org.apache.cassandra.distributed.shared.jmx.RMICloseableServerSocketFactory;
 import org.apache.cassandra.distributed.shared.JMXUtil;
 import org.apache.cassandra.utils.JMXServerUtils;
 import org.apache.cassandra.utils.MBeanWrapper;
-import org.apache.cassandra.utils.RMICloseableClientSocketFactory;
-import org.apache.cassandra.utils.RMICloseableServerSocketFactory;
 import sun.rmi.transport.tcp.TCPEndpoint;
 
 import static org.apache.cassandra.config.CassandraRelevantProperties.JAVA_RMI_DGC_LEASE_VALUE_IN_JVM_DTEST;
